examples: add userspace key provisioning using the TEE Client API
'hse-symkey-provision' uses the TEE Client API to invoke the optee core
and import the symmetric key provided in a file through a payload
structure.

The hse-symkey-provision program takes 3 arguments: a file containing
the symmetric key and its meta-data (the payload), the group where the
symmetric key will be imported and the slot from the same group. The
bytes read from the file are transmitted unaltered to the optee core.

'pta_hse_kp' is the interface exposed by optee_os' PTA. The file must be
kept in sync with optee_os' lib/libutee/include/pta_hse_kp.h.

Issue: ALB-9773
Signed-off-by: Bogdan Roman <bogdan-gabriel.roman@nxp.com>