examples: 'make all' builds only default examples
Add a new list of examples, default_examples, which will be built upon
calling make all. This change is needed for examples which have
dependencies to external components (i.e. TEE Client API) or special
build needs and do not rely on the default set of build parameters,
libraries, etc.
Mainly, new examples will be added to the examples list and should also
be included in the default_examples list if they're to be built by
default.
Add trustzone-key-provisioning only to examples list.

Issue: ALB-9773
Signed-off-by: Bogdan Roman <bogdan-gabriel.roman@nxp.com>