examples: avoid bus errors when comiling with -fstore-merging
The optimization flag 'store-merging', which is part of the O2
optimizations, merges contiguous stores of immediate values narrower
than a word into fewer wider stores to reduce the number of instructions
However, this results in unaligned accesses for the packed structures
of HSE when consecutive fields are assigned the same value.

In hse-secboot.c there are two such instances where unaligned accesses
are generated, resulting in bus errors at runtime. Workaround this
shortcoming by declaring the affected variables as 'volatile'.
Issue: ALB-9906

Signed-off-by: Bogdan Roman <bogdan-gabriel.roman@nxp.com>